Combined FpsCamera and FpsMovement into FpsController
Removed PlayerController
