+++ Assets/Scripts/FpsController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b6efc4df435d1ba46ada36b441c375da
+guid: 780fe7b9966b388438df8eb48b128a66
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
