--- Assets/Scripts/FpsMovement.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 48d542ddbab80a643a9164f91794a8f0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
